refactorización winner. Refactorizacion Game completa
